keep Normalizer
